+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - TODO
+
+### Added
+
+- In `events` argument, on event return value is sent back to Python
+- Internal: we now reference the echarts instance so anyone forking the project is able to call it back for further actions.
+
 ## [0.3.0] - 2021-04-16
 
 ### Added